chore: refactor test code to share common methods
