+++ core/src/avm1/globals.rs
@@ -117,14 +117,8 @@ pub fn create_globals<'gc>(gc_context: MutationContext<'gc, '_>) -> Object<'gc> 
 #[allow(clippy::unreadable_literal)]
 mod tests {
     use super::*;
-    use crate::avm1::activation::Activation;
+    use crate::avm1::test_utils::with_avm;
     use crate::avm1::Error;
-    use crate::backend::audio::NullAudioBackend;
-    use crate::backend::navigator::NullNavigatorBackend;
-    use crate::display_object::DisplayObject;
-    use crate::movie_clip::MovieClip;
-    use gc_arena::rootless_arena;
-    use rand::{rngs::SmallRng, SeedableRng};
 
     macro_rules! test_std {
         ( $test: ident, $fun: expr, $version: expr, $($args: expr => $out: expr),* ) => {
@@ -140,40 +134,6 @@ mod tests {
                 })
             }
         };
-    }
-
-    fn with_avm<F, R>(swf_version: u8, test: F) -> R
-    where
-        F: for<'a, 'gc> FnOnce(
-            &mut Avm1<'gc>,
-            &mut ActionContext<'a, 'gc, '_>,
-            GcCell<'gc, Object<'gc>>,
-        ) -> R,
-    {
-        rootless_arena(|gc_context| {
-            let mut avm = Avm1::new(gc_context);
-            let movie_clip: Box<dyn DisplayObject> = Box::new(MovieClip::new(gc_context));
-            let root = GcCell::allocate(gc_context, movie_clip);
-            let mut context = ActionContext {
-                gc_context,
-                global_time: 0,
-                root,
-                start_clip: root,
-                active_clip: root,
-                target_clip: Some(root),
-                target_path: Value::Undefined,
-                rng: &mut SmallRng::from_seed([0u8; 16]),
-                audio: &mut NullAudioBackend::new(),
-                navigator: &mut NullNavigatorBackend::new(),
-            };
-
-            let globals = avm.global_object_cell();
-            avm.insert_stack_frame(Activation::from_nothing(swf_version, globals, gc_context));
-
-            let this = root.read().object().as_object().unwrap().to_owned();
-
-            test(&mut avm, &mut context, this)
-        })
     }
 
     test_std!(boolean_function, boolean, 19,
