+++ core/src/avm1/globals/math.rs
@@ -121,20 +121,14 @@ pub fn create<'gc>(gc_context: MutationContext<'gc, '_>) -> GcCell<'gc, Object<'
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::avm1::activation::Activation;
+    use crate::avm1::test_utils::with_avm;
     use crate::avm1::Error;
-    use crate::backend::audio::NullAudioBackend;
-    use crate::backend::navigator::NullNavigatorBackend;
-    use crate::display_object::DisplayObject;
-    use crate::movie_clip::MovieClip;
-    use gc_arena::rootless_arena;
-    use rand::{rngs::SmallRng, SeedableRng};
 
     macro_rules! test_std {
         ( $test: ident, $name: expr, $($args: expr => $out: expr),* ) => {
             #[test]
             fn $test() -> Result<(), Error> {
-                with_avm(19, |avm, context| {
+                with_avm(19, |avm, context, _root| {
                     let math = create(context.gc_context);
                     let function = math.read().get($name, avm, context, math);
 
@@ -146,39 +140,6 @@ mod tests {
                 })
             }
         };
-    }
-
-    fn with_avm<F, R>(swf_version: u8, test: F) -> R
-    where
-        F: for<'a, 'gc> FnOnce(&mut Avm1<'gc>, &mut ActionContext<'a, 'gc, '_>) -> R,
-    {
-        rootless_arena(|gc_context| {
-            let mut avm = Avm1::new(gc_context, swf_version);
-            let movie_clip: Box<dyn DisplayObject> =
-                Box::new(MovieClip::new(swf_version, gc_context));
-            let root = GcCell::allocate(gc_context, movie_clip);
-            let mut context = ActionContext {
-                gc_context,
-                global_time: 0,
-                player_version: 32,
-                root,
-                start_clip: root,
-                active_clip: root,
-                target_clip: Some(root),
-                target_path: Value::Undefined,
-                rng: &mut SmallRng::from_seed([0u8; 16]),
-                audio: &mut NullAudioBackend::new(),
-                navigator: &mut NullNavigatorBackend::new(),
-            };
-
-            let globals = avm.global_object_cell();
-            avm.insert_stack_frame(
-                Activation::from_nothing(swf_version, globals, gc_context),
-                &mut context,
-            );
-
-            test(&mut avm, &mut context)
-        })
     }
 
     test_std!(test_abs, "abs",
@@ -268,7 +229,7 @@ mod tests {
 
     #[test]
     fn test_atan2_nan() {
-        with_avm(19, |avm, context| {
+        with_avm(19, |avm, context, _root| {
             let math = GcCell::allocate(context.gc_context, create(context.gc_context));
             assert_eq!(atan2(avm, context, *math.read(), &[]), Value::Number(NAN));
             assert_eq!(
@@ -303,7 +264,7 @@ mod tests {
 
     #[test]
     fn test_atan2_valid() {
-        with_avm(19, |avm, context| {
+        with_avm(19, |avm, context, _root| {
             let math = GcCell::allocate(context.gc_context, create(context.gc_context));
             assert_eq!(
                 atan2(avm, context, *math.read(), &[Value::Number(10.0)]),
